requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
streamlit>=1.29.0
typer>=0.9.0
pandas>=2.1.3
numpy>=1.26.2
scikit-learn>=1.3.2
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.18.0
python-multipart>=0.0.6
openpyxl>=3.1.2
pydantic>=2.5.2
pytest>=7.4.3
httpx>=0.25.2
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
joblib>=1.3.2
tqdm>=4.66.1
